requirements.txt
# Requirements for Djangular-Cli.
regex
prompt_toolkit
pip~=20.2.2
virtualenv~=20.0.30
setuptools~=49.3.1
django-environ # environ
distlib~=0.3.1
pyfiglet~=0.8.post1
Pygments~=2.6.1